tests: Move `fakeGetUserMedia` to `utils.js` for reuse
