+++ static/tests/frontend/specs/setStream.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const {cartesian} = require('ep_webrtc/static/tests/frontend/utils');
+const {cartesian, fakeGetUserMedia} = require('ep_webrtc/static/tests/frontend/utils');
 
 describe('setStream()', function () {
   let chrome$;
@@ -8,36 +8,6 @@ describe('setStream()', function () {
   const otherVideoId = `video_${otherUserId.replace(/\./g, '_')}`;
   const otherInterfaceId = `interface_${otherVideoId}`;
   let ownUserId, ownVideoId, ownInterfaceId;
-
-  const makeSilentAudioTrack = () => {
-    const ctx = new AudioContext();
-    const oscillator = ctx.createOscillator();
-    const dst = oscillator.connect(ctx.createMediaStreamDestination());
-    oscillator.start();
-    return dst.stream.getAudioTracks()[0];
-  };
-
-  const makeVideoTrack = () => {
-    const canvas = helper.padChrome$.window.document.createElement('canvas');
-    canvas.width = 160;
-    canvas.height = 120;
-    const ctx = canvas.getContext('2d');
-    ctx.fillStyle = `#${Math.floor(Math.random() * 2 ** 24).toString(16).padStart(6, '0')}`;
-    ctx.fillRect(0, 0, canvas.width, canvas.height);
-    return canvas.captureStream().getVideoTracks()[0];
-  };
-
-  // Creates dummy audio and/or video tracks. Limitations:
-  //   - `audio` and `video` are treated as Booleans (video size requirements are ignored).
-  //   - Most browsers prohibit audio until there has been some user interaction with the page or
-  //     the real getUserMedia() has been called.
-  const fakeGetUserMedia = async ({audio, video}) => {
-    if (!audio && !video) throw new DOMException('either audio or video is required', 'TypeError');
-    return new MediaStream([
-      ...(audio ? [makeSilentAudioTrack()] : []),
-      ...(video ? [makeVideoTrack()] : []),
-    ]);
-  };
 
   describe('Audio and video enabled', function () {
     const testCases = [...cartesian(...Array(4).fill([false, true]))].map(
